[tests] Renamed base test classes for consistency
Having files like test_admin.py is preferable over admin.py
because admin.py may be confused with the model admin definitions
during development.
