+++ django_netjsongraph/tests/base/test_admin.py
