+++ django_netjsongraph/tests/base/test_node.py
