+++ django_netjsongraph/tests/test_snapshot.py
@@ -2,7 +2,7 @@ from django.test import TestCase
 
 from . import CreateGraphObjectsMixin
 from ..models import Snapshot, Topology
-from .base.snapshot import TestSnapshotMixin
+from .base.test_snapshot import TestSnapshotMixin
 
 
 class TestSnapshot(TestSnapshotMixin, CreateGraphObjectsMixin, TestCase):
